# docs/swagger.yaml
basePath: /api
definitions:
  dto.AddMemberRequest:
    properties:
      role:
        enum:
        - owner
        - admin
        - member
        - viewer
        example: member
        type: string
      user_id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
    required:
    - role
    - user_id
    type: object
...
  dto.CreateOrgRequest:
    properties:
      name:
        example: Acme
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.ErrorEnvelope:
    properties:
      error:
        type: string
    type: object
...
  dto.MemberEnvelope:
    properties:
      member:
        $ref: '#/definitions/models.Membership'
      message:
        type: string
    type: object
  dto.MessageEnvelope:
    properties:
      message:
        type: string
    type: object
  dto.NotificationEnvelope:
    properties:
      message:
        type: string
      notification:
        $ref: '#/definitions/models.Notification'
    type: object
  dto.NotificationsEnvelope:
    properties:
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.PageMeta'
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
    type: object
  dto.OrgEnvelope:
    properties:
      message:
        type: string
      org:
        $ref: '#/definitions/models.Organization'
    type: object
  dto.OrgsEnvelope:
    properties:
      message:
        type: string
      orgs:
        items:
          $ref: '#/definitions/models.Organization'
        type: array
    type: object
  dto.PageMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.UnreadEnvelope:
    properties:
      message:
        type: string
      unread:
        example: 3
        type: integer
    type: object
...
  dto.UpdateMemberRequest:
    properties:
      role:
        enum:
        - owner
        - admin
        - member
        - viewer
        example: admin
        type: string
    required:
    - role
    type: object
...
  models.Membership:
    properties:
      created_at:
        type: string
      org_id:
        type: string
      role:
        example: admin
        type: string
      user_id:
        type: string
    type: object
  models.Notification:
    properties:
      body:
        example: Your email is verified. Glad to have you!
        type: string
      created_at:
        type: string
      data:
        additionalProperties: {}
        type: object
      id:
        type: string
      read_at:
        type: string
      title:
        example: Welcome to Grip
        type: string
      type:
        example: welcome
        type: string
      user_id:
        type: string
    type: object
  models.Organization:
    properties:
      created_at:
        type: string
      id:
        example: b3c4d5e6-f7a8-49b0-c1d2-e3f4a5b6c7d8
        type: string
      name:
        example: Acme
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Grip API with JWT auth and email verification.
  title: Grip API
  version: "1.0"
paths:
...
  /notifications/:
    get:
      parameters:
      - description: Unread only
        in: query
        name: unread_only
        type: boolean
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        example: 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationsEnvelope'
      security:
      - BearerAuth: []
      summary: List my notifications (newest first)
      tags:
      - notifications
  /notifications/{id}/read:
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NotificationEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Mark one notification read
      tags:
      - notifications
  /notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Mark all my notifications read
      tags:
      - notifications
  /notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UnreadEnvelope'
      security:
      - BearerAuth: []
      summary: Count my unread notifications
      tags:
      - notifications
  /orgs/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrgsEnvelope'
      security:
      - BearerAuth: []
      summary: List my organizations
      tags:
      - orgs
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrgRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.OrgEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create an organization (caller becomes owner)
      tags:
      - orgs
  /orgs/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Delete the organization
      tags:
      - orgs
  /orgs/{id}/members:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MemberEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a member to the organization
      tags:
      - orgs
  /orgs/{id}/members/{uid}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member user ID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Remove a member from the organization
      tags:
      - orgs
    put:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member user ID
        in: path
        name: uid
        required: true
        type: string
      - description: Role payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MemberEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Change a member's role
      tags:
      - orgs
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT. Example: "Bearer
      eyJhbGciOiJIUzI1NiJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"